Update 02-the-basics.md
note: "For this book, we will focus on using Chrome and its developer tools for two reasons: Chrome has a a set of versatile and powerful tools, and focusing on the tools of one browser helps keep the instructions simple." you say this yet two paragraphs away you tell the reader they should also look at Firefox! maybe reword that paragraph from "You should" to "It's also good to familiarize yourself with", etc.

Line 69, I changed from "We learned that", which doesn't make sense with line 70 ("We learned that: the javascript console". Reword?)